updated version with comments
